Refactor optimizer tests to test_optimizers
