+++ tests/models/test_gpu.py
@@ -87,55 +87,6 @@ def test_ddp_all_dataloaders_passed_to_fit(tmpdir):
     assert result == 1, "DDP doesn't work with dataloaders passed to fit()."
 
 
-def test_optimizer_return_options():
-    tutils.reset_seed()
-
-    trainer = Trainer()
-    model, hparams = tutils.get_default_model()
-
-    # single optimizer
-    opt_a = torch.optim.Adam(model.parameters(), lr=0.002)
-    opt_b = torch.optim.SGD(model.parameters(), lr=0.002)
-
-    model.configure_optimizers = lambda: opt_a
-    optim, lr_sched = trainer.init_optimizers(model)
-    assert len(optim) == 1 and len(lr_sched) == 0
-
-    # opt tuple
-    model.configure_optimizers = lambda: (opt_a, opt_b)
-    optim, lr_sched = trainer.init_optimizers(model)
-    assert len(optim) == 2 and optim[0] == opt_a and optim[1] == opt_b
-    assert len(lr_sched) == 0
-
-    # opt list
-    model.configure_optimizers = lambda: [opt_a, opt_b]
-    optim, lr_sched = trainer.init_optimizers(model)
-    assert len(optim) == 2 and optim[0] == opt_a and optim[1] == opt_b
-    assert len(lr_sched) == 0
-
-    # opt tuple of lists
-    scheduler = torch.optim.lr_scheduler.StepLR(opt_a, 10)
-    model.configure_optimizers = lambda: ([opt_a], [scheduler])
-    optim, lr_sched = trainer.init_optimizers(model)
-    assert len(optim) == 1 and len(lr_sched) == 1
-    assert optim[0] == opt_a and \
-        lr_sched[0] == dict(scheduler=scheduler, interval='epoch',
-                            frequency=1, reduce_on_plateau=False,
-                            monitor='val_loss')
-
-
-def test_default_optimizer_warning():
-    tutils.reset_seed()
-
-    trainer = Trainer()
-    model, hparams = tutils.get_default_model()
-
-    model.configure_optimizers = lambda: None
-
-    with pytest.warns(UserWarning, match='Adam optimizer with `lr=1e-3`'):
-        _, __ = trainer.init_optimizers(model)
-
-
 def test_cpu_slurm_save_load(tmpdir):
     """Verify model save/load/checkpoint on CPU."""
     tutils.reset_seed()
